requirements.txt
schedule
pylast
spotipy
python-dotenv
